feat(offline): Implement local map caching and download feature
This update allows the app to download and cache maps from supported online sources. Now, when a map gets selected for the first time, it gets downloaded and stored for offline use. Improved handling of Blob Storage and associated metadata allows easier access and management of local map data. Also, an option to remove downloaded maps was added.